requirements.txt
streamlit
google-generativeai
langchain
langchain_community
langchain_core
